Tree async changes support
(cherry picked from commit 713eeb8ea34635302a8daa0c85a9d2b9fd44a757)

+++ SpaceProfilerLogicTests/TestHelpers/FileSystemHelper.cs
@@ -27,6 +27,12 @@ public class FileSystemHelper
 
     public void Delete(string path)
     {
+        if (path == Root)
+        {
+            Clear();
+            return;
+        }
+            
         var fullPath = Path.GetFullPath(path, Root);
         if (File.Exists(fullPath))
             File.Delete(fullPath);
